+++ src/main/java/poopiemeow/TaskList.java
@@ -1,0 +1,36 @@
+package poopiemeow;
+
+import java.util.ArrayList;
+import poopiemeow.task.Task;
+
+public class TaskList {
+    private ArrayList<Task> tasks;
+
+    public TaskList(ArrayList<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public TaskList() {
+        this.tasks = new ArrayList<>();
+    }
+
+    public void addTask(Task task) {
+        tasks.add(task);
+    }
+
+    public Task deleteTask(int index) {
+        return tasks.remove(index);
+    }
+
+    public Task getTask(int index) {
+        return tasks.get(index);
+    }
+
+    public ArrayList<Task> getTasks() {
+        return tasks;
+    }
+
+    public int size() {
+        return tasks.size();
+    }
+}
