+++ src/main/java/poopiemeow/exception/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package poopiemeow.exception;
+
+public class EmptyDescriptionException extends Exception {
+    public EmptyDescriptionException(String message) {
+        super(message);
+    }
+}
